chore (package): Specify Node version and document versions.
Developers can unintentionally install and build with Node or `pnpm`
versions we don't officially support. Currently our CI runs with Node
v18 and v20, and we're using `pnpm` v8.

This change adds an `engines` field to the top level `package.json`
file. This will output a warning on install/build if the engine
version doesn't match the specified range.

The `pnpm` version is already specified in `packageManager` to reflect
the version we use in CI, although it is not enforced or warned if,
for example, the developer is on v9.

This change also updates the README to note supported Node engine
versions and `pnpm` version used.

We can consider for the future but choose not to pursue for now:
- placing `engine-strict=true` in the top level `.npmrc` to
  hard-fail on incompatible engine versions
- adding a preinstall/prebuild command that checks versions with
  more explicit failures, messaging, etc.